Added implementation planning/discovery documents
These documents can by used by administrators to audit their current instance of Salesforce, gather requirements, plan their communications, and determine the level of effort that will be required to implement the Interactions for Student Recruitment package.
 
**Project Plan and Timeline**
This spreadsheet will help administrators or project managers create a plan for implementing this package. 
 
**Communications and Training Plan**
This template is a starting point to help administrators create a plan to promote an implementation of the package internally and schedule training for end users. 
 
**Discovery Document**
This document was made to guide system administrators through some of the preliminary steps of auditing a current instance of Salesforce to determine the level of effort to install the Interactions package. It includes a starting set of discovery questions linked to the document sections administrators should review to prepare for the discovery.
 
**Interactions Discovery Document Template**
This spreadsheet takes the starting set of discovery questions outlined in the Discovery Document and organizes them to provide a place for administrators to record their findings. Questions are sorted by topic and include the reason each question is being asked.